requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests   # to call HF Inference API
openai>=1.40.0